Right, I saw that the javascript library for drawing the stats chart was closing in on 0.5MB, so decided that it needed to change. It was downloading the library clientside and building (using json data) clientside. Why couldn't I build the chart at site-generation time?
Thankfully, there's a different library for that :-)

However, it has required a huge reshuffle.
I've moved the chart into its own page (/statistics), produced from an .ejs file so I could write purer javascript than you get in .njk files.
Then the /stats page pulls it in via an iframe.

Then, because the chart page needs access to the stylesheet, and therefore the manifest.json, I've extracted manifest information out of eleventy and into a data file (/src/data/manifest.js). This allows it to be referenced in .eleventy and directly within the .ejs.

1. .eleventy re-working relationship to manifest.json into a require()
2. adding /src/data/manifest.js
3. rolling webpack.config back as stats entry no longer required (as stats chart no longer built clientside)
4. remove src/view/stats.ejs which produced /stats.json, as that's no longer required clientside
5. told postcss.config.js to look through .ejs files as well, because there's some css in those too!
6. cleaned out styles for the javascript chart.
6b. delete javascript chart-generating script.
7. declaring content going into meta-git tags as safe, so it isn't unnecessarily escaped.
8. neatening up sitemap output
9. package.json: adding `eleventy-charts` to build chart at site-generation time, and removing javascript charting libraries.
10. removing background images (banners) from pages, even though thumbnails are still used somewhere
11. new src/view/statistics.ejs to produce the chart as svg
12. updating stats.njk to pull in via iframe